[Persistence] Hide dangling underscores
Hide dangling underscores; JSLint considers these a problem
and the command line build fails as a consequence when they
are present directly in the code, but CouchDB provides these
so we cannot avoid them entirely. So, hide them behind
variables and use []-style lookup.

Concludes transition work for the CouchDB persistence adapter,
WTD-537.