[Layout] Add auto-save behavior
Add an EditRepresenter which detects changes to domain
object model or view configuration and, when these
occur, triggers mutation and persistence of the domain
object. (Note that in Edit mode, 'persistence' only
flags an object as dirty; actual persistence does not
occur until the user chooses Save.)

This supports editing of position/size in Layout mode,
and the persistence of those changes. WTD-535